requirements.txt
langchain
langchain-anthropic
pybit
python-dotenv
pydantic
pandas
numpy
